requirements.txt
discord.py==1.7.3[voice,speed]
aiohttp>=3.8.0
pillow>=10.0.0
ijson>=3.2.0
motor>=3.3.0
pymongo>=4.5.0
unidecode>=1.3.6
watchdog>=3.0.0
deepl>=1.15.0
aiofiles>=23.0.0
curl-cffi>=0.6.0
google-generativeai>=0.3.0
